requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
aiofiles=22.1.0=py310hecd8cb5_0
aiosqlite=0.18.0=py310hecd8cb5_0
anyio=3.5.0=py310hecd8cb5_0
appnope=0.1.2=py310hecd8cb5_1001
argon2-cffi=20.1.0=py310hca72f7f_1
asttokens=2.0.5=pyhd3eb1b0_0
attrs=23.1.0=py310hecd8cb5_0
babel=2.11.0=py310hecd8cb5_0
beautifulsoup4=4.12.2=py310hecd8cb5_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
bottleneck=1.3.5=py310h4e76f89_0
brotli=1.0.9=hca72f7f_7
brotli-bin=1.0.9=hca72f7f_7
brotlipy=0.7.0=py310hca72f7f_1002
bzip2=1.0.8=h1de35cc_0
ca-certificates=2023.12.12=hecd8cb5_0
certifi=2024.2.2=py310hecd8cb5_0
cffi=1.15.1=py310h6c40b1e_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
comm=0.1.2=py310hecd8cb5_0
contourpy=1.0.5=py310haf03e11_0
cryptography=38.0.4=py310hf6deb26_0
cycler=0.11.0=pyhd3eb1b0_0
debugpy=1.6.7=py310hcec6c5f_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.4=py310hecd8cb5_0
exceptiongroup=1.2.0=py310hecd8cb5_0
executing=0.8.3=pyhd3eb1b0_0
ffmpeg=4.3=h0a44026_0
fftw=3.3.9=h9ed2024_1
filelock=3.13.1=py310hecd8cb5_0
flit-core=3.6.0=pyhd3eb1b0_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=hd8bbffd_0
gettext=0.21.0=h7535e17_0
giflib=5.2.1=haf1e3a3_0
glib=2.69.1=hfff2838_2
gmp=6.2.1=he9d5cce_3
gmpy2=2.1.2=py310hd5de756_0
gnutls=3.6.15=hed9c0bf_0
gst-plugins-base=1.14.1=hcec6c5f_1
gstreamer=1.14.1=h6c40b1e_1
icu=58.2=h0a44026_3
idna=3.4=py310hecd8cb5_0
imbalanced-learn=0.10.1=pyhd8ed1ab_0
intel-openmp=2021.4.0=hecd8cb5_3538
ipykernel=6.25.0=py310h20db666_0
ipython=8.20.0=py310hecd8cb5_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=8.0.4=py310hecd8cb5_0
jedi=0.18.1=py310hecd8cb5_1
jinja2=3.1.3=py310hecd8cb5_0
joblib=1.1.1=py310hecd8cb5_0
jpeg=9e=hca72f7f_0
json5=0.9.6=pyhd3eb1b0_0
jsonschema=4.19.2=py310hecd8cb5_0
jsonschema-specifications=2023.7.1=py310hecd8cb5_0
jupyter=1.0.0=py310hecd8cb5_8
jupyter_client=7.4.9=py310hecd8cb5_0
jupyter_console=6.6.3=py310hecd8cb5_0
jupyter_core=5.5.0=py310hecd8cb5_0
jupyter_events=0.8.0=py310hecd8cb5_0
jupyter_server=1.23.4=py310hecd8cb5_0
jupyter_server_fileid=0.9.0=py310hecd8cb5_0
jupyter_server_ydoc=0.8.0=py310hecd8cb5_1
jupyter_ydoc=0.2.4=py310hecd8cb5_0
jupyterlab=3.6.3=py310hecd8cb5_0
jupyterlab_pygments=0.2.2=py310hecd8cb5_0
jupyterlab_server=2.22.0=py310hecd8cb5_0
jupyterlab_widgets=3.0.9=py310hecd8cb5_0
kiwisolver=1.4.4=py310hcec6c5f_0
krb5=1.20.1=hdba6334_1
lame=3.100=h1de35cc_0
lcms2=2.12=hf1fd2bf_0
lerc=3.0=he9d5cce_0
libbrotlicommon=1.0.9=hca72f7f_7
libbrotlidec=1.0.9=hca72f7f_7
libbrotlienc=1.0.9=hca72f7f_7
libclang=12.0.0=default_hbc2896b_2
libcxx=14.0.6=h9765a3e_0
libdeflate=1.8=h9ed2024_5
libedit=3.1.20230828=h6c40b1e_0
libffi=3.4.2=hecd8cb5_6
libgfortran=5.0.0=11_3_0_hecd8cb5_28
libgfortran5=11.3.0=h9dfd629_28
libiconv=1.16=hca72f7f_2
libidn2=2.3.2=h9ed2024_0
libjpeg-turbo=2.0.0=hca72f7f_0
libllvm12=12.0.0=h9b2ccf5_3
libpng=1.6.37=ha441bb4_0
libpq=12.15=hdb2fb19_1
libsodium=1.0.18=h1de35cc_0
libtasn1=4.16.0=h9ed2024_0
libtiff=4.5.0=h2cd0358_0
libunistring=0.9.10=h9ed2024_0
libwebp=1.2.4=h56c3ce4_0
libwebp-base=1.2.4=hca72f7f_0
libxml2=2.9.14=hbf8cd5e_0
llvm-openmp=14.0.6=h0dcd299_0
lz4-c=1.9.4=hcec6c5f_0
markupsafe=2.1.3=py310h6c40b1e_0
matplotlib=3.6.2=py310hecd8cb5_0
matplotlib-base=3.6.2=py310h220de94_0
matplotlib-inline=0.1.6=py310hecd8cb5_0
mistune=2.0.4=py310hecd8cb5_0
mkl=2021.4.0=hecd8cb5_637
mkl-service=2.4.0=py310hca72f7f_0
mkl_fft=1.3.1=py310hf879493_0
mkl_random=1.2.2=py310hc081a56_0
mpc=1.1.0=h6ef4df4_1
mpfr=4.0.2=h9066e36_1
mpmath=1.3.0=py310hecd8cb5_0
munkres=1.1.4=py_0
nbclassic=1.0.0=py310hecd8cb5_0
nbclient=0.8.0=py310hecd8cb5_0
nbconvert=7.10.0=py310hecd8cb5_0
nbformat=5.9.2=py310hecd8cb5_0
ncurses=6.4=hcec6c5f_0
nest-asyncio=1.5.6=py310hecd8cb5_0
nettle=3.7.3=h230ac6f_1
networkx=3.1=py310hecd8cb5_0
notebook=6.5.4=py310hecd8cb5_1
notebook-shim=0.2.3=py310hecd8cb5_0
nspr=4.35=hcec6c5f_0
nss=3.89.1=hcec6c5f_0
numexpr=2.8.4=py310h9638375_0
numpy=1.23.5=py310h9638375_0
numpy-base=1.23.5=py310ha98c3c9_0
openh264=2.1.1=h8346a28_0
openssl=1.1.1w=hca72f7f_0
packaging=22.0=py310hecd8cb5_0
pandas=1.5.2=py310h3ea8b11_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
patsy=0.5.2=py310hecd8cb5_1
pcre=8.45=h23ab428_0
pexpect=4.8.0=pyhd3eb1b0_3
pillow=9.3.0=py310h81888ad_1
pip=22.3.1=py310hecd8cb5_0
platformdirs=3.10.0=py310hecd8cb5_0
ply=3.11=py310hecd8cb5_0
prometheus_client=0.14.1=py310hecd8cb5_0
prompt-toolkit=3.0.43=py310hecd8cb5_0
prompt_toolkit=3.0.43=hd3eb1b0_0
psutil=5.9.0=py310hca72f7f_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.15.1=py310hecd8cb5_1
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py310hecd8cb5_0
pyqt=5.15.10=py310hcec6c5f_0
pyqt5-sip=12.13.0=py310h6c40b1e_0
pysocks=1.7.1=py310hecd8cb5_0
python=3.10.8=h218abb5_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.16.2=py310hecd8cb5_0
python-json-logger=2.0.7=py310hecd8cb5_0
pytorch=2.2.0=py3.10_0
pytz=2022.7=py310hecd8cb5_0
pyyaml=6.0.1=py310h6c40b1e_0
pyzmq=23.2.0=py310he9d5cce_0
qt-main=5.15.2=h719ae48_7
qtconsole=5.5.0=py310hecd8cb5_0
qtpy=2.4.1=py310hecd8cb5_0
readline=8.2=hca72f7f_0
referencing=0.30.2=py310hecd8cb5_0
requests=2.28.1=py310hecd8cb5_0
rfc3339-validator=0.1.4=py310hecd8cb5_0
rfc3986-validator=0.1.1=py310hecd8cb5_0
rpds-py=0.10.6=py310hf2ad997_0
scikit-learn=1.1.3=py310he9d5cce_0
scipy=1.9.3=py310h09290a1_0
seaborn=0.12.2=py310hecd8cb5_0
send2trash=1.8.2=py310hecd8cb5_0
setuptools=65.6.3=py310hecd8cb5_0
sip=6.7.12=py310hcec6c5f_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.3.0=py310hecd8cb5_0
soupsieve=2.5=py310hecd8cb5_0
sqlite=3.41.2=h6c40b1e_0
stack_data=0.2.0=pyhd3eb1b0_0
statsmodels=0.13.2=py310hca72f7f_0
sympy=1.12=py310hecd8cb5_0
terminado=0.17.1=py310hecd8cb5_0
threadpoolctl=2.2.0=pyh0d69192_0
tinycss2=1.2.1=py310hecd8cb5_0
tk=8.6.12=h5d9f67b_0
tomli=2.0.1=py310hecd8cb5_0
torchaudio=2.2.0=py310_cpu
torchmetrics=0.11.0=pyhd8ed1ab_0
torchvision=0.17.0=py310_cpu
tornado=6.2=py310hca72f7f_0
traitlets=5.7.1=py310hecd8cb5_0
typing_extensions=4.4.0=py310hecd8cb5_0
tzdata=2022g=h04d1e81_0
urllib3=1.26.13=py310hecd8cb5_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py310hecd8cb5_1
websocket-client=0.58.0=py310hecd8cb5_4
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=4.0.5=py310hecd8cb5_0
xz=5.2.8=h6c40b1e_0
y-py=0.5.9=py310h7242b5c_0
yaml=0.2.5=haf1e3a3_0
ypy-websocket=0.8.2=py310hecd8cb5_0
zeromq=4.3.5=hcec6c5f_0
zlib=1.2.13=h4dc903c_0
zstd=1.5.2=hcb37349_0
